+++ source/common/upstream/health_checker_base_impl.cc
@@ -1,6 +1,6 @@
 #include "common/upstream/health_checker_base_impl.h"
 
-#include "envoy/api/v2/core/health_check_logging.pb.h"
+#include "envoy/data/core/v2alpha/health_check_event.pb.h"
 
 #include "common/router/router.h"
 
@@ -163,7 +163,7 @@ void HealthCheckerImplBase::setUnhealthyCrossThread(const HostSharedPtr& host) {
       return;
     }
 
-    session->second->setUnhealthy(envoy::api::v2::core::HealthCheckFailureType::PASSIVE);
+    session->second->setUnhealthy(envoy::data::core::v2alpha::HealthCheckFailureType::PASSIVE);
   });
 }
 
@@ -204,8 +204,7 @@ void HealthCheckerImplBase::ActiveHealthCheckSession::handleSuccess() {
       parent_.incHealthy();
       changed_state = HealthTransition::Changed;
       if (parent_.event_logger_) {
-        parent_.event_logger_->logAddHealthy(parent_.healthCheckerType(), host_,
-                                             parent_.healthy_threshold_, first_check_);
+        parent_.event_logger_->logAddHealthy(parent_.healthCheckerType(), host_, first_check_);
       }
     } else {
       changed_state = HealthTransition::ChangePending;
@@ -221,20 +220,19 @@ void HealthCheckerImplBase::ActiveHealthCheckSession::handleSuccess() {
 }
 
 HealthTransition HealthCheckerImplBase::ActiveHealthCheckSession::setUnhealthy(
-    envoy::api::v2::core::HealthCheckFailureType type) {
+    envoy::data::core::v2alpha::HealthCheckFailureType type) {
   // If we are unhealthy, reset the # of healthy to zero.
   num_healthy_ = 0;
 
   HealthTransition changed_state = HealthTransition::Unchanged;
   if (!host_->healthFlagGet(Host::HealthFlag::FAILED_ACTIVE_HC)) {
-    if (type != envoy::api::v2::core::HealthCheckFailureType::NETWORK ||
+    if (type != envoy::data::core::v2alpha::HealthCheckFailureType::NETWORK ||
         ++num_unhealthy_ == parent_.unhealthy_threshold_) {
       host_->healthFlagSet(Host::HealthFlag::FAILED_ACTIVE_HC);
       parent_.decHealthy();
       changed_state = HealthTransition::Changed;
       if (parent_.event_logger_) {
-        parent_.event_logger_->logEjectUnhealthy(parent_.healthCheckerType(), host_, type,
-                                                 parent_.timeout_, parent_.unhealthy_threshold_);
+        parent_.event_logger_->logEjectUnhealthy(parent_.healthCheckerType(), host_, type);
       }
     } else {
       changed_state = HealthTransition::ChangePending;
@@ -242,9 +240,9 @@ HealthTransition HealthCheckerImplBase::ActiveHealthCheckSession::setUnhealthy(
   }
 
   parent_.stats_.failure_.inc();
-  if (type == envoy::api::v2::core::HealthCheckFailureType::NETWORK) {
+  if (type == envoy::data::core::v2alpha::HealthCheckFailureType::NETWORK) {
     parent_.stats_.network_failure_.inc();
-  } else if (type == envoy::api::v2::core::HealthCheckFailureType::PASSIVE) {
+  } else if (type == envoy::data::core::v2alpha::HealthCheckFailureType::PASSIVE) {
     parent_.stats_.passive_failure_.inc();
   }
 
@@ -254,7 +252,7 @@ HealthTransition HealthCheckerImplBase::ActiveHealthCheckSession::setUnhealthy(
 }
 
 void HealthCheckerImplBase::ActiveHealthCheckSession::handleFailure(
-    envoy::api::v2::core::HealthCheckFailureType type) {
+    envoy::data::core::v2alpha::HealthCheckFailureType type) {
   HealthTransition changed_state = setUnhealthy(type);
   timeout_timer_->disableTimer();
   interval_timer_->enableTimer(parent_.interval(HealthState::Unhealthy, changed_state));
@@ -268,23 +266,18 @@ void HealthCheckerImplBase::ActiveHealthCheckSession::onIntervalBase() {
 
 void HealthCheckerImplBase::ActiveHealthCheckSession::onTimeoutBase() {
   onTimeout();
-  handleFailure(envoy::api::v2::core::HealthCheckFailureType::NETWORK);
+  handleFailure(envoy::data::core::v2alpha::HealthCheckFailureType::NETWORK);
 }
 
 void HealthCheckEventLoggerImpl::logEjectUnhealthy(
-    envoy::api::v2::core::HealthCheckerType health_checker_type,
+    envoy::data::core::v2alpha::HealthCheckerType health_checker_type,
     const HostDescriptionConstSharedPtr& host,
-    envoy::api::v2::core::HealthCheckFailureType failure_type, std::chrono::milliseconds timeout,
-    uint32_t unhealthy_threshold) {
-  envoy::api::v2::core::HealthCheckEvent event;
+    envoy::data::core::v2alpha::HealthCheckFailureType failure_type) {
+  envoy::data::core::v2alpha::HealthCheckEvent event;
   event.set_health_checker_type(health_checker_type);
   event.set_host_address(host->address()->asString());
   event.set_cluster_name(host->cluster().name());
   event.mutable_eject_unhealthy_event()->set_failure_type(failure_type);
-  event.mutable_eject_unhealthy_event()->mutable_timeout()->set_nanos(
-      std::chrono::duration_cast<std::chrono::nanoseconds>(timeout).count());
-  event.mutable_eject_unhealthy_event()->mutable_unhealthy_threshold()->set_value(
-      unhealthy_threshold);
   // Make sure the type enums make it into the JSON
   const auto json = MessageUtil::getJsonStringFromMessage(event, /* pretty_print */ false,
                                                           /* always_print_primitive_fields */ true);
@@ -292,13 +285,12 @@ void HealthCheckEventLoggerImpl::logEjectUnhealthy(
 }
 
 void HealthCheckEventLoggerImpl::logAddHealthy(
-    envoy::api::v2::core::HealthCheckerType health_checker_type,
-    const HostDescriptionConstSharedPtr& host, uint32_t healthy_threshold, bool first_check) {
-  envoy::api::v2::core::HealthCheckEvent event;
+    envoy::data::core::v2alpha::HealthCheckerType health_checker_type,
+    const HostDescriptionConstSharedPtr& host, bool first_check) {
+  envoy::data::core::v2alpha::HealthCheckEvent event;
   event.set_health_checker_type(health_checker_type);
   event.set_host_address(host->address()->asString());
   event.set_cluster_name(host->cluster().name());
-  event.mutable_add_healthy_event()->mutable_healthy_threshold()->set_value(healthy_threshold);
   event.mutable_add_healthy_event()->set_first_check(first_check);
   // Make sure the type enums make it into the JSON
   const auto json = MessageUtil::getJsonStringFromMessage(event, /* pretty_print */ false,